move uniqueQueryId into its own file
